moved contents of /http/userToken into /userToken for consistency and updated references
